// src/shared/models/CardInfo.ts
export default interface CardInfo
{
    id: string;
    name: string;
    skill: string;
    grade: number;
    power: number;
    shield: number;
    ability: string;
    nation: string;
    race: string;
    sets: string[];
    keywords: string[];
    image: string;
    type: string;
    critical: number;
    trigger: string;
}